requirements.txt
# Chỉ cài CPU version của PyTorch để giảm size và thời gian build
torch==2.0.1+cpu
torchvision==0.15.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Streamlit và dependencies
streamlit==1.28.0
Pillow==10.0.0
matplotlib==3.7.2